Core / Preparation Added new form @html Added new function to create cells Added new function to count cells

+++ timeFrame.go
@@ -4,6 +4,6 @@ package main
 type TimeFrame struct {
 	ID                      uint32
 	Value                   float64
-	IncludingMinute         uint32
+	IncludingMinute         float64
 	AboveBelowCellThreshold bool
 }
